moved csrf protection to journalist_app/__init__.py
